day 20 rust refactor

+++ Rust/day_20/src/main.rs
@@ -1,4 +1,4 @@
-const DECRYPTION_KEY: usize = 811589153;
+const DECRYPTION_KEY: i64 = 811589153;
 
 fn parse_input(lines: &str) -> Vec<(usize, i64)> {
     lines
@@ -17,8 +17,9 @@ fn solve(mut lines: Vec<(usize, i64)>, rounds: usize) -> i64 {
             if n.1 == 0 {
                 continue;
             }
+
             let new_index = ((old_index as i64 + n.1).rem_euclid(length as i64 - 1)) as usize;
-            let old  = lines.remove(old_index);
+            let old = lines.remove(old_index);
             lines.insert(new_index, old);
         }
     }
@@ -30,13 +31,13 @@ fn solve(mut lines: Vec<(usize, i64)>, rounds: usize) -> i64 {
 }
 
 fn part1(lines: Vec<(usize, i64)>) -> i64 {
-    solve( lines, 1)
+    solve(lines, 1)
 }
 
 fn part2(lines: Vec<(usize, i64)>) -> i64 {
     let lines = lines
         .into_iter()
-        .map(|(index, value)| (index, value * DECRYPTION_KEY as i64))
+        .map(|(index, value)| (index, value * DECRYPTION_KEY))
         .collect();
     solve(lines, 10)
 }
